requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
enum-compat==0.0.2
eventlet==0.20.1
Flask==0.12
Flask-Admin==1.4.2
Flask-Login==0.4.0
Flask-SocketIO==2.8.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
greenlet==0.4.12
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pbr==1.10.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-engineio==1.1.2
python-socketio==1.6.3
rauth==0.7.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
WTForms==2.1
